% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{A character string.}
}
\value{
A character string greeting.
}
\description{
A very small example function.
}
\examples{
hello("Ralph")
}
